+++ web-common/src/features/dashboards/dashboard-queries.ts
@@ -1,4 +1,4 @@
-import { isMetricsViewQuery } from "@rilldata/web-local/lib/svelte-query/invalidation";
+import { isMetricsViewQuery } from "@rilldata/web-common/runtime-client/invalidation";
 import type { QueryClient } from "@tanstack/svelte-query";
 
 export function cancelDashboardQueries(
