+++ web-common/src/features/entity-management/actions.ts
@@ -6,11 +6,11 @@ import type {
   V1RenameFileAndReconcileResponse,
 } from "@rilldata/web-common/runtime-client";
 import { httpRequestQueue } from "@rilldata/web-common/runtime-client/http-client";
-import type { ActiveEntity } from "@rilldata/web-local/lib/application-state-stores/app-store";
 import {
   invalidateAfterReconcile,
   removeEntityQueries,
-} from "@rilldata/web-local/lib/svelte-query/invalidation";
+} from "@rilldata/web-common/runtime-client/invalidation";
+import type { ActiveEntity } from "@rilldata/web-local/lib/application-state-stores/app-store";
 import type {
   CreateBaseMutationResult,
   QueryClient,
