+++ web-common/src/features/entity-management/sync-file-system.ts
@@ -13,10 +13,10 @@ import {
 } from "@rilldata/web-common/runtime-client";
 import type { Page } from "@sveltejs/kit";
 import type { QueryClient } from "@tanstack/svelte-query";
-import { get, Readable, Writable } from "svelte/store";
+import { Readable, Writable, get } from "svelte/store";
 import type { FeatureFlags } from "../../../../web-local/src/lib/application-state-stores/application-store";
-import { invalidateAfterReconcile } from "../../../../web-local/src/lib/svelte-query/invalidation";
 import { overlay } from "../../layout/overlay-store";
+import { invalidateAfterReconcile } from "../../runtime-client/invalidation";
 import type { Runtime } from "../../runtime-client/runtime-store";
 import { getFilePathFromPagePath } from "./entity-mappers";
 import {
