+++ web-common/src/features/sources/add-source/createSource.ts
@@ -7,7 +7,7 @@ import type {
   V1PutFileAndReconcileResponse,
   V1ReconcileError,
 } from "@rilldata/web-common/runtime-client";
-import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
+import { invalidateAfterReconcile } from "@rilldata/web-common/runtime-client/invalidation";
 import type {
   CreateBaseMutationResult,
   QueryClient,
